tests: parametrized loaders with supported classes/models
